requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
